Divide header as component
